+++ Client/package.json
@@ -14,14 +14,16 @@
     "@htilssu/wowo": "^1.1.8",
     "@mantine/core": "^7.12.2",
     "@mantine/hooks": "^7.12.2",
+    "@react-buddy/ide-toolbox": "^2.4.0",
     "@react-oauth/google": "^0.12.1",
     "framer-motion": "^11.5.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-facebook-login": "^4.1.1",
+    "react-github-login": "^1.0.3",
     "react-icons": "^5.3.0",
-    "react-toastify": "^10.0.5",
-    "@react-buddy/ide-toolbox": "^2.4.0"
+    "react-oauth": "^0.0.1",
+    "react-toastify": "^10.0.5"
   },
   "devDependencies": {
     "@eslint/js": "^9.10.0",
