+++ Client/src/routes/routeConfig.jsx
@@ -1,12 +1,10 @@
 import {createBrowserRouter} from 'react-router-dom';
-
 import Home from "../pages/HomePage.jsx";
 import {MainLayout} from "../layouts/MainLayout.jsx";
 import {PageNotFound} from "../pages/PageNotFound.jsx";
-import SignIn from '../modules/core/components/account/SignIn.jsx';
-import Test from '../modules/core/components/account/test.jsx';
 import SignUp from './../pages/sign-up/SignUp';
 import ForgotPassword from './../pages/forgot-password/ForgotPassword';
+import SignIn from '../pages/sign-in/SignIn.jsx';
 
 
 
@@ -20,9 +18,6 @@ export const router = createBrowserRouter([
           },
           {
             path: 'sign-in', element: <SignIn/>,
-          },
-          {
-            path: 'test', element: <Test/>,
           },
         ],
         errorElement: <PageNotFound/>,
